requirements.txt
altgraph @ file:///C:/ci_311/altgraph_1676434977885/work
certifi==2024.8.30
charset-normalizer==3.4.0
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pefile @ file:///C:/b/abs_68oy9ka6fm/croot/pefile_1720651782540/work
pillow==11.0.0
pyinstaller @ file:///C:/b/abs_09h007ff52/croot/pyinstaller_1721682848440/work
pyinstaller-hooks-contrib @ file:///C:/b/abs_deumjuw0e0/croot/pyinstaller-hooks-contrib_1721675898298/work
pystray==0.19.5
pywin32-ctypes @ file:///C:/b/abs_2cfx5l4nvi/croot/pywin32-ctypes_1709340246092/work
requests==2.32.3
six==1.16.0
urllib3==2.2.3